+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.2.2-pre
 * Add result timestamp (when check is started)
+* Rename http(s) attributes: statusCodes -> statusCode, texts -> text
+* All checker attributes now support <attribute> and <attribute>-or
 
 ### 0.2.1
 * Add file checker